Support new template deployer version
